Fix execution of installed scripts
Python's method of resolving the filepaths of imported modules subtly
differs based on the nature of the importing file. If the importing file
is itself a Python module, the resolution works as expected. If the
importing file is executed as a script, the rules change.

That change in module resolution prevented the Python scripts, when
installed, from importin chadv_dev_tools.py.

The best way to fix the problem is (1) to install chadv_dev_tools.py into
$(prefix)/lib/python3.X/site-packages as a proper Python package and (2)
patch the scripts during installation to teach them installed location
of chadv_dev_tools.py. Sadly, this fix requires rewriting and moving
many of the source and make files.

Take extra care to ensure that the scripts continue to work if ran
directly from the source directory (that is, if they're not installed).