README: In installation walkthrough, fix PREFIX entanglement
The prefix into which the dev tools are installed is distinct from the
prefix into which the graphics libraries are installed. The dev tools
prefix is unique, but the graphics libraries may be installed multiple
times into multiple independent prefixes.

Clarifying the above point required rewriting so much of the walkthrough
that I took the opportunity to clarify a few other things too.